backend/requirements.txt
fastapi
uvicorn
sqlalchemy
geopandas
pandas
numpy
scikit-learn
pysal
shapely
osmnx
psycopg2-binary
geopy
rtree
h3
python-dotenv
pydantic
requests
aiohttp
pymysql
sqlalchemy_utils